Remove 'base' machine
This is no longer needed, it was an awkward workaround from when I was
worse at this and could get the "real" configuration booting in a VM,
that works now though. (I can't remember what the exact issue was).

Because I am still setting the hostname in aethelred.nix, and also just
so that things all keep working the same for me, I'll leave "aethelred"
in the name for now.